Add files via upload
The Test_student code is not my own work, I do not claim ownership. I have added this so that when examining my code you can test for all the collected functions which I have created in the three files for my coursework in Scientific Computation.

Inshallah, it can benefit me in the future when looking back or even anyone examining this work.